updated for new layout, removed graphs, removed caching
caching deprecated: source no longer provides graphs, this was the main reason for caching; it took time to load in, selenium would have to sit and wait for it. this is no longer the case. now we can return a result in sub 3 seconds. everything required is commented out, may be readded in future.